updated purchase test with page class and objects
