+++ features/purchase.feature
@@ -6,6 +6,7 @@ Feature: Purchase Feature
   Scenario:  Validate successful purchase text
     Then I will login as 'standard_user'
     And I will add the backpack to the cart
+    And I go to the cart
     And I go to checkout
     And I fill in the form details
     And I select Finish
