+++ pages/product.page.ts
@@ -23,29 +23,12 @@ export class Product {
         await this.page.locator(this.addToCart).click()
     }
 
-    public async goToCheckout() {
+    public async goToCart() {
         await this.page.locator(this.cartLink).click()
-        await this.page.waitForLoadState('networkidle')
-        await this.page.click(this.checkoutButton)
-        await this.page.waitForLoadState('networkidle')
     }
 
-    public async fillInCheckoutDetails(firstName: string, lastName: string, postalCode: string) {
-        await this.page.fill(this.firstNameField, firstName)
-        await this.page.fill(this.lastNameField, lastName)
-        await this.page.fill(this.postalCodeField, postalCode)
-        await this.page.click(this.continueButton)
-    }
-
-    public async finishCheckout() {
-        await this.page.click(this.finishButton)
-    }
-
-    public async validateText(expectedText: string) {
-        const actualText = await this.page.textContent(this.completeHeader);
-        if (actualText !== expectedText) {
-            throw new Error(`Expected text to be "${expectedText}" but found "${actualText}"`);
-        }
+    public async goToCheckout() {
+        await this.page.click(this.checkoutButton);
     }
 
     public async sortItemsBy(sortOption: string) {
