+++ steps/product.steps.ts
@@ -1,9 +1,14 @@
 import { Then, When } from '@cucumber/cucumber';
 import { getPage } from '../playwrightUtilities';
 import { Product } from '../pages/product.page';
+import { Checkout } from '../pages/checkout.page';
 
 Then('I will add the backpack to the cart', async () => {
   await new Product(getPage()).addBackPackToCart();
+});
+
+Then('I go to the cart', async () => {
+  await new Product(getPage()).goToCart();
 });
 
 Then('I go to checkout', async () => {
@@ -11,15 +16,15 @@ Then('I go to checkout', async () => {
 });
 
 Then('I fill in the form details', async () => {
-  await new Product(getPage()).fillInCheckoutDetails('John', 'Doe', '12345');
+  await new Checkout(getPage()).fillInCheckoutDetails('John', 'Doe', '12345');
 });
 
 Then('I select Finish', async () => {
-  await new Product(getPage()).finishCheckout();
+  await new Checkout(getPage()).finishCheckout();
 });
 
 Then('I validate the text {string}', async (expectedText) => {
-  await new Product(getPage()).validateText(expectedText);
+  await new Checkout(getPage()).validateText(expectedText);
 });
 When('I sort the items by {string}', async (sortOption) => {
   await new Product(getPage()).sortItemsBy(sortOption);
